investment_bot/requirements.txt
pandas
numpy
yfinance
ccxt
ta
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
python-dotenv

